refactor: break down sidebar shell into sub components
